refactor: Use Go types in ax25_link.go and more
While I'm trying to keep these refactors quite isolated into small-ish
commits, ax25_link / ax25_pad2 / dlq / server make enough calls to each
other that it felt reasonable to do a bunch of changes across them
together to save adding a bunch of extra temporary compatibility code.

+++ src/ptt.go
@@ -1048,7 +1048,7 @@ func ptt_set_real(ot int, channel int, ptt_signal int) {
 	 */
 
 	// #ifndef TEST
-	dlq_channel_busy(C.int(channel), C.int(ot), C.int(ptt_signal))
+	dlq_channel_busy(channel, C.int(ot), C.int(ptt_signal))
 	// #endif
 
 	/*
